fix(models): widen Unix timestamp fields from int to long (Y2K38)

+++ src/Corsinvest.ProxmoxVE.Api.Shared/Models/Cluster/ClusterBackup.cs
@@ -110,10 +110,10 @@ public class ClusterBackup : ModelBase, IStorageItem
     public string NotesTemplate { get; set; }
 
     /// <summary>
-    /// NextRun
+    /// NextRun (seconds since epoch)
     /// </summary>
     [JsonProperty("next-run")]
-    public int NextRun { get; set; }
+    public long NextRun { get; set; }
 
     /// <summary>
     /// Node
