+++ src/Corsinvest.ProxmoxVE.Api.Shared/Models/Cluster/ClusterLog.cs
@@ -73,7 +73,7 @@ public class ClusterLog : ModelBase
     /// </summary>
     [JsonProperty("time")]
     [DisplayFormat(DataFormatString = FormatHelper.DataFormatUnixTime)]
-    public int Time { get; set; }
+    public long Time { get; set; }
 
     /// <summary>
     /// Time
